Fix the tests for diff positions.
